@@ -2,27 +2,27 @@
   "scripts": {
     "clean": "rm -rf ./public/bundle.*",
     "build": "NODE_ENV=development webpack --mode development",
-    "build:prod": "NODE_ENV=production webpack -p"
+    "build:prod": "NODE_ENV=production webpack --mode production"
   },
   "dependencies": {
-    "firebase": "^9.1.3",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1"
+    "firebase": "^10.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "ajv": "^6.12.6",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-plugin-transform-class-properties": "^6.24.1",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
+    "@babel/cli": "^7.23.4",
+    "@babel/core": "^7.23.3",
+    "babel-loader": "^9.1.3",
+    "@babel/plugin-transform-class-properties": "^7.23.3",
+    "@babel/preset-env": "^7.23.3",
+    "@babel/preset-react": "^7.23.3",
     "babel-register": "^6.26.0",
-    "css-loader": "^0.28.11",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "style-loader": "^0.21.0",
-    "webpack": "^4.44.2",
-    "webpack-cli": "^3.3.12"
+    "css-loader": "^6.8.1",
+    "mini-css-extract-plugin": "^2.7.6",
+    "firebase-tools": "^12.9.1",
+    "style-loader": "^3.3.3",
+    "webpack": "^5.89.0",
+    "webpack-cli": "^5.1.4"
   },
   "private": true
 }
